requirements.txt
numpy
torch
torchvision
torchsummary
pandas
Pillow
matplotlib
seaborn
scikit-learn
pytest